requirements.txt
python-decouple>=3.3
pycoingecko>=1.4.1